[Course Groups] Refactoring of the course groups teams connector to make use of direct team creation and to change the procedure of how teams are created. Restoring old teams that have previously been removed is no longer possible.

+++ src/Chamilo/Libraries/Protocol/Microsoft/Graph/Service/GroupService.php
@@ -201,12 +201,19 @@ class GroupService
      *
      * @throws GroupNotExistsException
      * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
+     * @throws AzureUserNotExistsException
      */
     public function removeMemberFromGroup($groupId, User $user)
     {
         if ($this->isMemberOfGroup($groupId, $user))
         {
             $azureUserIdentifier = $this->getAzureUserIdentifier($user);
+
+            if (empty($azureUserIdentifier))
+            {
+                throw new AzureUserNotExistsException($user);
+            }
+
             $this->getGroupRepository()->removeMemberFromGroup($groupId, $azureUserIdentifier);
         }
     }
